rubber rats
crazy? i was crazy once they locked me crazy once they locked me in a room. a room. a rubber room a rubber room. a rubber room. a rubber room. a rubber room a room. a room. a rubber room. a rubber room. a rubber room with rats make me crazy i was crazy once they locked me in a room. a room a rubber room. a rubber room with rats make me crazy i was crazy once they locked me crazy i was crazy i was crazy once they locked me in a rubber room with rats. and rats make me in a rubber room. a rubber room a rubber room with rats. and rats and rats make me in a rubber room. a room. a rubber room